Prevent starvation

+++ src/bq_common/platform/thread/spin_lock.h
@@ -220,7 +220,7 @@ namespace bq {
 
         public:
             spin_lock_rw_crazy()
-                : state_(0)
+                : state_(static_cast<uint64_t>(0))
             {
             }
